+++ py_src/jupyter_lsp/trait_types.py
@@ -1,3 +1,4 @@
+import six
 import traitlets
 
 
@@ -20,3 +21,21 @@ class Schema(traitlets.Any):
                 )
             )
         return value
+
+
+class LoadableCallable(traitlets.TraitType):
+    """A trait which (maybe) loads a callable."""
+
+    info_text = "a loadable callable"
+
+    def validate(self, obj, value):
+        if isinstance(value, str):
+            try:
+                value = traitlets.import_item(value)
+            except Exception:
+                self.error(obj, value)
+
+        if six.callable(value):
+            return value
+        else:
+            self.error(obj, value)